Use the RC SDK to build if you're on RC
This allows build to succeed at installing the extensions, so that it's
possible to debug changes.

The wrinkles here are that:
1. SDK's before the RC one don't know how to install on RC.
2. The RC SDK *requires* a <Prerequisites> element
3. The Willow installer fails to install vsixes that have a <Prerequisites> element.

To workaround 2 and 3, we are *not* adding the element yet, but and we're
suppressing the requirement for it by turning off manifest schema validation.

Once 3 is fixed, we'll add the <Prerequisites> and remove the bypass.